Fixed bug and cleaned up skyread a bit

+++ sketches/skyread/setup.hpp
@@ -3,4 +3,4 @@
 int resistors[] = {2, 5, 4, 3}; // resistor pins
 int motors[] = {9, 10};         // motor pins
 int pos[] = {0, 0};             // initial position of motors
-int forever = 1e10;             // when finished delay forever
+long int forever = 1e10;        // when finished delay forever
